Improves bin monitoring intervals and battery handling
Reduces log and bin polling intervals for faster updates and simulates gradual battery drain for more realistic testing. Adds detailed debug output for fill level calculations and updates battery to use floating point precision. Temporarily disables server-side log generation.

+++ arduino/sBin/globals.h
@@ -70,7 +70,7 @@ extern long wifiTime;
 extern float lat;
 extern float lng;
 extern int fillLevel;
-extern int battery;
+extern float battery;
 extern String health;
 
 // ===== Timeouts =====
